Introduce busy waiting based on TIM3
Created very simple delay mechanism based on TIM3.
In order to make fixed time delay we run the counter and wait
for overrun interrupt in kind of busy wait, but with WFI
which is hopefully less power consuming than simple
looping.

HD44780 code now uses the delay in display reset.